requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pycryptodome>=3.21.0
python-dotenv>=1.0.0
apscheduler>=3.10.4
pytest>=8.0.0
